+++ src/components/Insight/InsightCardInternals.js
@@ -5,8 +5,8 @@ import moment from 'moment'
 import InsightTags from './InsightTags'
 import ProfileInfo from './ProfileInfo'
 import MultilineText from '../MultilineText/MultilineText'
+import LikeBtn from '../Like/LikeBtn'
 import styles from './InsightCard.module.scss'
-import InsightCardLikeBtn from './InsightCardLikeBtn'
 
 const InsightCardInternals = ({
   id,
@@ -43,7 +43,7 @@ const InsightCardInternals = ({
           />
         </div>
         <div className={styles.right}>
-          <InsightCardLikeBtn
+          <LikeBtn
             likesNumber={totalVotes}
             liked={!!votedAt}
             onClick={onLike}
