+++ src/components/Insight/insightsGQL.js
@@ -42,19 +42,3 @@ export const ALL_INSIGHTS_QUERY = gql`
     }
   }
 `
-
-export const UNLIKE_INSIGHT_MUTATION = gql`
-  mutation unvote($id: Int!) {
-    unvote(postId: $id) {
-      votedAt
-    }
-  }
-`
-
-export const LIKE_INSIGHT_MUTATION = gql`
-  mutation vote($id: Int!) {
-    vote(postId: $id) {
-      votedAt
-    }
-  }
-`
