+++ src/components/Like/LikeBtn.module.scss
@@ -1,0 +1,20 @@
+@import "~@santiment-network/ui/mixins.scss";
+
+.wrapper {
+    @include text-normal;
+    display: inline-flex;
+    align-items: center;
+    cursor: pointer;
+    line-height: 20px;
+    color: var(--black);
+    fill: var(--grey-dark);
+}
+
+.icon {
+    margin-right: 6px;
+}
+
+.liked {
+    color: var(--green);
+    fill: var(--green);
+}
