+++ src/components/Like/actions.js
@@ -1,0 +1,2 @@
+export const INSIGHTS_LIKE = '[voting] INSIGHTS_LIKE'
+export const INSIGHT_LIKE_FAIL = '[voting] INSIGHTS_LIKE_FAIL'
